+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## New Features
 
+- **_[Crosshair] Health/Ammo Bars_** setting
 - **_Translucent [Weapon] When Invisible_** setting
 - **NUGHUD Weapon icon**
 - **Pressing _Place [Automap] Mark_ while over a mark now changes that mark's color**